refactor: split the CLI into separate modules
